Anagram
We use Java 8 streams to collect the strings into groups of anagrams based on their sorted forms. The groupingBy collector is used to group the strings.

We then filter the groups to keep only those with more than one string using the filter operation.

Finally, we use the forEach operation to print the anagram groups.